Core: move AutoPatchRegister next to AutoWorld and AutoLogicRegister
