+++ worlds/sm/Rom.py
@@ -3,7 +3,8 @@ import os
 
 import json
 import Utils
-from Patch import read_rom, APDeltaPatch
+from Patch import read_rom
+from worlds.Files import APDeltaPatch
 
 SMJUHASH = '21f3e98df4780ee1c667b84e57d88675'
 ROM_PLAYER_LIMIT = 65535 # max archipelago player ID. note, SM ROM itself will only store 201 names+ids max
